caesiumcoin
A cryptocurrency (or crypto currency) is a digital asset designed to work as a medium of exchange that uses cryptography to secure its transactions, to control the creation of additional units, and to verify the transfer of assets. Cryptocurrencies are classified as a subset of digital currencies and are also classified as a subset of alternative currencies and virtual currencies.

Caesiumcoin, created in 2018, is decentralized cryptocurrency. Numerous other cryptocurrencies have been created. These are frequently called altcoins, as a blend of alternative coin. Caesiumcoin and its derivatives use decentralized control as opposed to centralized electronic money and central banking systems. The decentralized control is related to the use of caesiumcoin's blockchain transaction database in the role of a distributed ledger.